requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
flasgger==0.9.1
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.13.1
Flask-JWT-Simple==0.0.3
Flask-Login==0.4.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.3.0
nose==1.3.7
pep8==1.7.1
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
PyJWT==1.4.2
pylint==2.2.2
pytest==4.0.2
pytest-cov==2.6.1
PyYAML==3.13
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
